Revert "Merge branch 'master' into delta"
This reverts commit a1f13c0d140e58c69e0e9ab4bb0b217fc0e36738, reversing
changes made to ae5b80d49dd596d2e42e6ff82e2c562b739fb286.

+++ xbmc/cores/VideoPlayer/VideoRenderers/DebugRenderer.cpp
@@ -108,7 +108,7 @@ void CDebugRenderer::CRenderer::Render(int idx)
 {
   std::vector<COverlay*> render;
   std::vector<SElement>& list = m_buffers[idx];
-  float posY = 0.0f;
+  int posY = 0;
   for (std::vector<SElement>::iterator it = list.begin(); it != list.end(); ++it)
   {
     COverlay* o = nullptr;
